+++ NEWS.md
@@ -8,6 +8,7 @@
   - CAPI: GEOSHilbertCode (GH-556, Brendan Ward)
   - CAPI: GEOSGeom_createRectangle (GH-558, Brendan Ward)
   - CAPI: GEOSGeom_transformXY (GH-563, Dan Baston/Brendan Ward)
+  - CAPI: GEOSRemoveRepeatedPoints (GH-599, Paul Ramsey)
 
 - Fixes/Improvements:
   - Fix unaryUnion to avoid segfault with empty polygon (GH-501, Mike Taves)